pyproject.toml
[project]
name = "e-commerce-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "google-generativeai>=0.8.5",
    "ollama>=0.5.1",
    "openai>=1.97.0",
    "python-dotenv>=1.1.1",
    "requests==2.28.1",
    "streamlit>=1.47.0",
]
